Feat/kgms/v3.0.0/chenjiawei

See merge request kgcloud/kg-cloud-kgms!302

+++ kgsdk/src/main/java/com/plantdata/kgcloud/plantdata/req/explore/relation/RuleRelationGraphParameter.java
@@ -18,6 +18,6 @@ import java.util.Map;
 public class RuleRelationGraphParameter extends TimeRelationGraphParameter implements RuleGraphParameter {
 
     private List<RelationReasoningBean> rule;
-    private Map<Integer, JSONObject> reasoningRuleConfigs;
+    private Map<Long, JSONObject> reasoningRuleConfigs;
 
 }
